+++ matrix-sdk/src/main/java/org/matrix/androidsdk/data/store/MXMemoryStore.java
@@ -1637,11 +1637,6 @@ public class MXMemoryStore implements IMXStore {
     }
 
     @Override
-    public void setRoomsWithoutURLPreview(Set<String> roomIds) {
-        mMetadata.mRoomsListWithoutURLPrevew = roomIds;
-    }
-
-    @Override
     public void setUserWidgets(Map<String, Object> contentDict) {
         mMetadata.mUserWidgets = contentDict;
     }
@@ -1649,11 +1644,6 @@ public class MXMemoryStore implements IMXStore {
     @Override
     public Map<String, Object> getUserWidgets() {
         return mMetadata.mUserWidgets;
-    }
-
-    @Override
-    public Set<String> getRoomsWithoutURLPreviews() {
-        return (null != mMetadata.mRoomsListWithoutURLPrevew) ? mMetadata.mRoomsListWithoutURLPrevew : new HashSet<String>();
     }
 
     @Override