+++ CHANGES.rst
@@ -2,16 +2,21 @@ Changes to Matrix Android SDK in 0.9.34 (2020-XX-XX)
 =======================================================
 
 Features:
- -
+ - MSC2437: Store tagged events in Room Account Data
 
 Improvements:
+ - Enhance the room account data API naming.
  - MXSession: Do not refresh TURN servers when VoIP is not supported
 
 Bugfix:
  -
 
 API Change:
- -
+ - RoomAccountdata.hasTags() has been deprecated. Use .hasRoomTags() instead.
+ - RoomAccountdata.getKeys() has been deprecated. Use .getRoomTagsKeys() instead.
+ - RoomAccountdata.handleTagEvent() has been removed. Use .handleEvent() instead.
+ - IMXStore.setRoomsWithoutURLPreview() has been removed.
+ - IMXStore.getRoomsWithoutURLPreviews() has been removed. Use RoomAccountdata.isURLPreviewAllowedByUser() instead.
 
 Translations:
  -